Log message when scheme is illegal or not defined.
When the client connects to klishd the klishd tries to create
ksession. Firstly it gets scheme starting entry. When the starting
entry is not found it means that scheme is illegal or not defined.
Often klishd can't found scheme configuration. Previously log says
that session can't be created. It's true but doesn't explain the
possible cause of problem. Now the message is logged that says
about probably illegal scheme.