klishd: The daemon's listen socket can be chown-ed and chmod-ed.
The klishd config file can contain the following new options:

* UnixSocketUser
* UnixSocketGroup
* UnixSocketMode

The klishd daemon will chmod or/and chown created listen socket to the
specified values. If options are not specified the klishd will not perform
any chmod or chown.

Thanks to A. Eremin <korbezolympus@gmail.com> and
V. Chashchin <v.chashchin@nagtech.ru> for the original patch.